feat: replace sequences page with eval view #662
Move EvalView (richer matrix with filters, sort, view modes, and admin
support toggle) to /sequences as the primary listing experience for all
users. The admin-only /admin/eval route is removed; its ?user=email
deep link continues to work via /sequences?user=email.

- Sequences index renders EvalView with PageContainer wrapper and
  validated search params
- EvalView empty state now offers a "Create Your First Sequence" CTA
  for non-admin users with no sequences and no active filters
- /admin/usage support links and CSV supportUrl now point to /sequences
- Eval link removed from admin user dropdown
- Unused SequencesList component deleted
- routeTree.gen.ts pruned of /admin/eval entries (also auto-regenerates
  on next bun dev)
