--- src/routes/_protected/admin/eval.tsx
@@ -1,18 +1,0 @@
-import { EvalView } from '@/components/eval/eval-view';
-import { createFileRoute } from '@tanstack/react-router';
-import { z } from 'zod';
-
-const searchSchema = z.object({
-  user: z.string().email().optional(),
-});
-
-export const Route = createFileRoute('/_protected/admin/eval')({
-  validateSearch: searchSchema,
-  component: EvalPage,
-  staticData: { breadcrumb: 'Eval' },
-});
-
-function EvalPage() {
-  const { user } = Route.useSearch();
-  return <EvalView initialUserFilter={user} />;
-}
